requirements.txt
garminconnect>=0.2.13
pandas>=2.2.2
pyyaml>=6.0.2
requests>=2.32.3
pyserial>=3.5
bleak>=0.22.3